no uint admitted

+++ src/preprocess/add_feature.py
@@ -14,7 +14,7 @@ class PreprocessAddFeature(BaseFeature, PreprocessInit):
         self.base_data = self.base_data.with_columns(
             pl.col('WEEK_NUM')
             .alias(self.fold_time_col)
-            .cast(pl.UInt16)
+            .cast(pl.Int16)
         )
     
     def create_credit_bureau_a_1_feature(self) -> None:
@@ -503,14 +503,14 @@ class PreprocessAddFeature(BaseFeature, PreprocessInit):
             self.person_1.filter(
                 pl.col('num_group1')!=0
             ).group_by('case_id').agg(
-                pl.len().alias('number_rowsX').cast(pl.UInt8),
-                pl.col('childnum_185L').max().cast(pl.UInt8),
+                pl.len().alias('number_rowsX').cast(pl.Int16),
+                pl.col('childnum_185L').max().cast(pl.Int16),
                 *[
                     (
                         pl.col(column_name)
                         .filter(
                             (pl.col(column_name)==self.mapper_mask['person_1'][column_name][self.hashed_missing_label])
-                        ).count().alias(f'{column_name}_a55475b1countX').cast(pl.UInt8)
+                        ).count().alias(f'{column_name}_a55475b1countX').cast(pl.Int16)
                     )
                     for column_name in [
                         'contaddr_district_15M', 'contaddr_zipcode_807M', 
@@ -535,7 +535,7 @@ class PreprocessAddFeature(BaseFeature, PreprocessInit):
                         )
                         .count()
                         .alias(f'{column_name}_{col_value}X')
-                        .cast(pl.UInt8)
+                        .cast(pl.Int16)
                     )
                     for column_name, col_value in [
                         (column_name, col_value) 
@@ -551,7 +551,7 @@ class PreprocessAddFeature(BaseFeature, PreprocessInit):
                         )
                         .len()
                         .alias(f'{column_name}_nullX')
-                        .cast(pl.UInt8)
+                        .cast(pl.Int16)
                     )
                     for column_name in [
                         'gender_992L', 'housingtype_772L', 
@@ -618,11 +618,11 @@ class PreprocessAddFeature(BaseFeature, PreprocessInit):
                     pl.col('case_id').filter(
                         (pl.col('num_group1')!=0) &
                         (pl.col('num_group2')==0)                
-                    ).count().alias('related_n_0_X').cast(pl.UInt16),
+                    ).count().alias('related_n_0_X').cast(pl.Int16),
                     pl.col('case_id').filter(
                         (pl.col('num_group1')==0) &
                         (pl.col('num_group2')!=0)                
-                    ).count().alias('related_0_n_X').cast(pl.UInt16),   
+                    ).count().alias('related_0_n_X').cast(pl.Int16),   
                 ] +
                 [
                     (
@@ -633,7 +633,7 @@ class PreprocessAddFeature(BaseFeature, PreprocessInit):
                         )
                         .count()
                         .alias(f'{col[:-1]}_{single_value}_0_n_' + col[-1])
-                        .cast(pl.UInt16)
+                        .cast(pl.Int16)
                     )
                     for col, single_value in zero_n_list
                 ] + 
@@ -646,7 +646,7 @@ class PreprocessAddFeature(BaseFeature, PreprocessInit):
                         )
                         .count()
                         .alias(f'{col[:-1]}_{single_value}_n_0_' + col[-1])
-                        .cast(pl.UInt16)
+                        .cast(pl.Int16)
                     )
                     for col, single_value in n_0_list
                 ]
@@ -736,11 +736,11 @@ class PreprocessAddFeature(BaseFeature, PreprocessInit):
                     pl.col('case_id').filter(
                         (pl.col('num_group1')!=0) &
                         (pl.col('num_group2')==0)                
-                    ).count().alias('related_n_0_X').cast(pl.UInt16),
+                    ).count().alias('related_n_0_X').cast(pl.Int16),
                     pl.col('case_id').filter(
                         (pl.col('num_group1')==0) &
                         (pl.col('num_group2')!=0)                
-                    ).count().alias('related_0_n_X').cast(pl.UInt16),   
+                    ).count().alias('related_0_n_X').cast(pl.Int16),   
                 ] +
                 [
                     (
@@ -751,7 +751,7 @@ class PreprocessAddFeature(BaseFeature, PreprocessInit):
                         )
                         .count()
                         .alias(f'{col[:-1]}_{single_value}_0_n_' + col[-1])
-                        .cast(pl.UInt16)
+                        .cast(pl.Int16)
                     )
                     for col, single_value in category_list
                 ] + 
@@ -760,7 +760,7 @@ class PreprocessAddFeature(BaseFeature, PreprocessInit):
                         (pl.col(col)==self.mapper_mask['applprev_2'][col][single_value])&
                         (pl.col('num_group1')!=0) &
                         (pl.col('num_group2')==0)
-                    ).count().alias(f'{col[:-1]}_{single_value}_n_0_' + col[-1]).cast(pl.UInt16)
+                    ).count().alias(f'{col[:-1]}_{single_value}_n_0_' + col[-1]).cast(pl.Int16)
                     for col, single_value in category_list
                 ]
             )
